Committing using repo.postsync.d hook

+++ portage/media-sound/pulseaudio/pulseaudio-12.0-r1.ebuild
@@ -282,7 +282,9 @@ multilib_src_install_all() {
 		systemd_newtmpfilesd "${FILESDIR}/${PN}.tmpfiles" "${PN}.conf"
 	else
 		# Prevent warnings when system-wide is not used, bug #447694
-		rm "${ED%/}"/etc/dbus-1/system.d/pulseaudio-system.conf || die
+		if use dbus ; then
+			rm "${ED%/}"/etc/dbus-1/system.d/pulseaudio-system.conf || die
+		fi
 	fi
 
 	if use zeroconf ; then
@@ -309,13 +311,6 @@ pkg_postinst() {
 		elog "    https://www.freedesktop.org/wiki/Software/PulseAudio/Documentation/User/WhatIsWrongWithSystemWide/"
 		elog "    https://www.freedesktop.org/wiki/Software/PulseAudio/Documentation/User/SystemWide/"
 		elog "    https://wiki.gentoo.org/wiki/PulseAudio#Headless_server"
-		if use gnome ; then
-			elog
-			elog "By enabling gnome USE flag, you enabled gconf support. Please note"
-			elog "that you might need to remove the gnome USE flag or disable the"
-			elog "gconf module on /etc/pulse/system.pa to be able to use PulseAudio"
-			elog "with a system-wide instance."
-		fi
 	fi
 
 	if use equalizer && ! use qt5; then
